Accept encoding keyword in initialize
There appears to be no tests for this, but it was reported as
missing behavior in jruby/jruby#7478. This change accepts the
encoding keyword and sets it into the StringIO object, but uses
of that encoding are still missing. In CRuby calls to the get_enc
macro are used when reading from the StringIO, to encode the
resulting Strings. These changes will come in a future commit.