Add arity-checking anywhere it is missing (#48)
As part of jruby/jruby#7751 we are recommending that all extension code
manually check the arity of incoming arguments to variable- arity
methods, as in CRuby. This ensures that all call paths will be checked,
including direct paths from Java or invokedynamic, and avoids array
indexing errors in these situations.