Eagerly defrost chilled strings
Ref: https://bugs.ruby-lang.org/issues/20205

One minor backward compatibility issue with chilled strings
is that since they claim to be frozen, code using the pattern:

```ruby
StringIO.new("")
```

Is now producing read only `StringIO` instances. It would be
preferable to emit a warning instead.

We can detect whether the current Ruby produce chilled strings
and if so preemptively attempt to mutate the given string to
produce the warning immediately and preserve the original
behavior.