Add rb_enc_check that ignores empty str1
The version of this method in JRuby does not support passing an
encoding for the first argument, which means downstream code in
enc_compatible_latter will not reject cases StringIO is expected
to reject. This patch hacks a version of rb_enc_check that works
with current JRuby versions and exhibits the behavior required.

A future JRuby update will improve this API and this hack should
become a fallback at that point (for older JRuby versions).